User panel file cleanup
