+++ app/public/views/sessionErrorPanel.js
